+++ main-service/src/main/java/ru/practicum/user/service/UserService.java
@@ -10,5 +10,5 @@ public interface UserService {
 
     List<UserDto> getUsers(List<Long> ids, Integer from, Integer size);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 }
